% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gbox.R
\name{gbox}
\alias{gbox}
\title{Create a ggplot box plot.}
\usage{
gbox(data, x_var, y_var = NULL, orientation = "vertical")
}
\arguments{
\item{data}{A data frame containing the variables to be plotted.}

\item{x_var}{The variable for the x-axis.}

\item{y_var}{(Optional) The variable for the y-axis.}

\item{orientation}{The orientation of the box plots: "vertical" or "horizontal".}
}
\value{
This function returns a ggplot box plot object.
}
\description{
This function creates a ggplot-style box plot of a given numeric vector,
or side-by-side box plots if both a numeric and a categorical variable are provided.
}
\examples{
## Create a box plot of a numeric vector.
data <- data.frame(values = rnorm(100))
gbox(data, "values", orientation = "vertical")

## Create side-by-side box plots of a numeric and a categorical variable.
data <- data.frame(values = rnorm(100), category = sample(c("A", "B"), 100, replace = TRUE))
gbox(data, "values", "category", orientation = "vertical")

}
