% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghist.R
\name{ghist}
\alias{ghist}
\title{Create a ggplot histogram.}
\usage{
ghist(x, bins = 10)
}
\arguments{
\item{x}{A vector of numeric values to be plotted.}

\item{bins}{Number of bins for the histogram.}
}
\value{
This function returns a ggplot histogram object.
}
\description{
This function creates a ggplot-style histogram of a given vector.
}
\examples{
## Create a histogram of a numeric vector.
data <- rnorm(100)
ghist(data)

}
